# general-service/api/qubic/aggregation/general/v1/service.pb.gw.go
// Code generated by protoc-gen-grpc-gateway. DO NOT EDIT.
// source: qubic/aggregation/general/v1/service.proto

/*
Package v1 is a reverse proxy.

It translates gRPC into RESTful JSON APIs.
*/
package v1

import (
	"context"
	"errors"
	"io"
	"net/http"

	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
	"github.com/grpc-ecosystem/grpc-gateway/v2/utilities"
	"google.golang.org/grpc"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/grpclog"
	"google.golang.org/grpc/metadata"
	"google.golang.org/grpc/status"
	"google.golang.org/protobuf/proto"
)

// Suppress "imported and not used" errors
var (
	_ codes.Code
	_ io.Reader
	_ status.Status
	_ = errors.New
	_ = runtime.String
	_ = utilities.NewDoubleArray
	_ = metadata.Join
)

func request_AggregationGeneralService_GetCurrentIpoBids_0(ctx context.Context, marshaler runtime.Marshaler, client AggregationGeneralServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var (
		protoReq GetCurrentIpoBidsRequest
		metadata runtime.ServerMetadata
	)
	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && !errors.Is(err, io.EOF) {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}
	if req.Body != nil {
		_, _ = io.Copy(io.Discard, req.Body)
	}
	msg, err := client.GetCurrentIpoBids(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err
}

func local_request_AggregationGeneralService_GetCurrentIpoBids_0(ctx context.Context, marshaler runtime.Marshaler, server AggregationGeneralServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var (
		protoReq GetCurrentIpoBidsRequest
		metadata runtime.ServerMetadata
	)
	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && !errors.Is(err, io.EOF) {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}
	msg, err := server.GetCurrentIpoBids(ctx, &protoReq)
	return msg, metadata, err
}

func request_AggregationGeneralService_GetIdentitiesBalances_0(ctx context.Context, marshaler runtime.Marshaler, client AggregationGeneralServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var (
		protoReq GetIdentitiesBalancesRequest
		metadata runtime.ServerMetadata
	)
	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && !errors.Is(err, io.EOF) {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}
	if req.Body != nil {
		_, _ = io.Copy(io.Discard, req.Body)
	}
	msg, err := client.GetIdentitiesBalances(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err
}

func local_request_AggregationGeneralService_GetIdentitiesBalances_0(ctx context.Context, marshaler runtime.Marshaler, server AggregationGeneralServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var (
		protoReq GetIdentitiesBalancesRequest
		metadata runtime.ServerMetadata
	)
	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && !errors.Is(err, io.EOF) {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}
	msg, err := server.GetIdentitiesBalances(ctx, &protoReq)
	return msg, metadata, err
}

// RegisterAggregationGeneralServiceHandlerServer registers the http handlers for service AggregationGeneralService to "mux".
// UnaryRPC     :call AggregationGeneralServiceServer directly.
// StreamingRPC :currently unsupported pending https://github.com/grpc/grpc-go/issues/906.
// Note that using this registration option will cause many gRPC library features to stop working. Consider using RegisterAggregationGeneralServiceHandlerFromEndpoint instead.
// GRPC interceptors will not work for this type of registration. To use interceptors, you must use the "runtime.WithMiddlewares" option in the "runtime.NewServeMux" call.
func RegisterAggregationGeneralServiceHandlerServer(ctx context.Context, mux *runtime.ServeMux, server AggregationGeneralServiceServer) error {
	mux.Handle(http.MethodPost, pattern_AggregationGeneralService_GetCurrentIpoBids_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		annotatedContext, err := runtime.AnnotateIncomingContext(ctx, mux, req, "/qubic.aggregation.general.v1.AggregationGeneralService/GetCurrentIpoBids", runtime.WithHTTPPathPattern("/getCurrentIpoBids"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_AggregationGeneralService_GetCurrentIpoBids_0(annotatedContext, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}
		forward_AggregationGeneralService_GetCurrentIpoBids_0(annotatedContext, mux, outboundMarshaler, w, req, response_AggregationGeneralService_GetCurrentIpoBids_0{resp.(*GetCurrentIpoBidsResponse)}, mux.GetForwardResponseOptions()...)
	})
	mux.Handle(http.MethodPost, pattern_AggregationGeneralService_GetIdentitiesBalances_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		annotatedContext, err := runtime.AnnotateIncomingContext(ctx, mux, req, "/qubic.aggregation.general.v1.AggregationGeneralService/GetIdentitiesBalances", runtime.WithHTTPPathPattern("/getIdentitiesBalances"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_AggregationGeneralService_GetIdentitiesBalances_0(annotatedContext, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}
		forward_AggregationGeneralService_GetIdentitiesBalances_0(annotatedContext, mux, outboundMarshaler, w, req, response_AggregationGeneralService_GetIdentitiesBalances_0{resp.(*GetIdentitiesBalancesResponse)}, mux.GetForwardResponseOptions()...)
	})

	return nil
}

// RegisterAggregationGeneralServiceHandlerFromEndpoint is same as RegisterAggregationGeneralServiceHandler but
// automatically dials to "endpoint" and closes the connection when "ctx" gets done.
func RegisterAggregationGeneralServiceHandlerFromEndpoint(ctx context.Context, mux *runtime.ServeMux, endpoint string, opts []grpc.DialOption) (err error) {
	conn, err := grpc.NewClient(endpoint, opts...)
	if err != nil {
		return err
	}
	defer func() {
		if err != nil {
			if cerr := conn.Close(); cerr != nil {
				grpclog.Errorf("Failed to close conn to %s: %v", endpoint, cerr)
			}
			return
		}
		go func() {
			<-ctx.Done()
			if cerr := conn.Close(); cerr != nil {
				grpclog.Errorf("Failed to close conn to %s: %v", endpoint, cerr)
			}
		}()
	}()
	return RegisterAggregationGeneralServiceHandler(ctx, mux, conn)
}

// RegisterAggregationGeneralServiceHandler registers the http handlers for service AggregationGeneralService to "mux".
// The handlers forward requests to the grpc endpoint over "conn".
func RegisterAggregationGeneralServiceHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
	return RegisterAggregationGeneralServiceHandlerClient(ctx, mux, NewAggregationGeneralServiceClient(conn))
}

// RegisterAggregationGeneralServiceHandlerClient registers the http handlers for service AggregationGeneralService
// to "mux". The handlers forward requests to the grpc endpoint over the given implementation of "AggregationGeneralServiceClient".
// Note: the gRPC framework executes interceptors within the gRPC handler. If the passed in "AggregationGeneralServiceClient"
// doesn't go through the normal gRPC flow (creating a gRPC client etc.) then it will be up to the passed in
// "AggregationGeneralServiceClient" to call the correct interceptors. This client ignores the HTTP middlewares.
func RegisterAggregationGeneralServiceHandlerClient(ctx context.Context, mux *runtime.ServeMux, client AggregationGeneralServiceClient) error {
	mux.Handle(http.MethodPost, pattern_AggregationGeneralService_GetCurrentIpoBids_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		annotatedContext, err := runtime.AnnotateContext(ctx, mux, req, "/qubic.aggregation.general.v1.AggregationGeneralService/GetCurrentIpoBids", runtime.WithHTTPPathPattern("/getCurrentIpoBids"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_AggregationGeneralService_GetCurrentIpoBids_0(annotatedContext, inboundMarshaler, client, req, pathParams)
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}
		forward_AggregationGeneralService_GetCurrentIpoBids_0(annotatedContext, mux, outboundMarshaler, w, req, response_AggregationGeneralService_GetCurrentIpoBids_0{resp.(*GetCurrentIpoBidsResponse)}, mux.GetForwardResponseOptions()...)
	})
	mux.Handle(http.MethodPost, pattern_AggregationGeneralService_GetIdentitiesBalances_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		annotatedContext, err := runtime.AnnotateContext(ctx, mux, req, "/qubic.aggregation.general.v1.AggregationGeneralService/GetIdentitiesBalances", runtime.WithHTTPPathPattern("/getIdentitiesBalances"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_AggregationGeneralService_GetIdentitiesBalances_0(annotatedContext, inboundMarshaler, client, req, pathParams)
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}
		forward_AggregationGeneralService_GetIdentitiesBalances_0(annotatedContext, mux, outboundMarshaler, w, req, response_AggregationGeneralService_GetIdentitiesBalances_0{resp.(*GetIdentitiesBalancesResponse)}, mux.GetForwardResponseOptions()...)
	})
	return nil
}

type response_AggregationGeneralService_GetCurrentIpoBids_0 struct {
	*GetCurrentIpoBidsResponse
}

func (m response_AggregationGeneralService_GetCurrentIpoBids_0) XXX_ResponseBody() interface{} {
	response := m.GetCurrentIpoBidsResponse
	return response.IpoTransactions
}

type response_AggregationGeneralService_GetIdentitiesBalances_0 struct {
	*GetIdentitiesBalancesResponse
}

func (m response_AggregationGeneralService_GetIdentitiesBalances_0) XXX_ResponseBody() interface{} {
	response := m.GetIdentitiesBalancesResponse
	return response.Balances
}

var (
	pattern_AggregationGeneralService_GetCurrentIpoBids_0     = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0}, []string{"getCurrentIpoBids"}, ""))
	pattern_AggregationGeneralService_GetIdentitiesBalances_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0}, []string{"getIdentitiesBalances"}, ""))
)

var (
	forward_AggregationGeneralService_GetCurrentIpoBids_0     = runtime.ForwardResponseMessage
	forward_AggregationGeneralService_GetIdentitiesBalances_0 = runtime.ForwardResponseMessage
)